Fix the account of getEvents
