+++ BackEnd/ChaukaApp.ServiceAPI/Controllers/AccountsController.cs
@@ -1,5 +1,6 @@
 ﻿namespace ChaukaApp.ServiceAPI.Controllers
 {
+    using System;
     using System.Web.Http;
     using BusinessLogic;
     using DataAccess;
@@ -25,6 +26,27 @@
         {
             IMapPositioner positioner = new MapPositioner();
             var result = positioner.UpdateUserPosition(user);
+        }
+
+
+        [HttpGet]
+        [Route("api/v1/accounts/{userId}/events")]
+        public IResult<UserEvent> GetEventsByUser(int userId)
+        {
+            IResult<UserEvent> resultEvent = new ResultEvents();
+            IEventHost userVerifier = new EventHost();
+
+            try
+            {
+                resultEvent = userVerifier.GetUserEvents(userId);
+            }
+            catch (Exception)
+            {
+                resultEvent.Success = false;
+                resultEvent.Message = "The service could not respond to your request";
+            }
+
+            return resultEvent;
         }
     }
 }
