+++ BackEnd/ChaukaApp.ServiceAPI/Controllers/GuestsController.cs
@@ -44,25 +44,5 @@
 
             return result;
         }
-
-        [HttpGet]
-        [Route("api/v1/accounts/{userId}/events")]
-        public IResult<UserEvent> GetEventsByUser(int userId)
-        {
-            IResult<UserEvent> resultEvent = new ResultEvents();
-            IEventHost userVerifier = new EventHost();
-
-            try
-            {
-                resultEvent = userVerifier.GetUserEvents(userId);
-            }
-            catch (Exception)
-            {
-                resultEvent.Success = false;
-                resultEvent.Message = "The service could not respond to your request";
-            }
-
-            return resultEvent;
-        }
     }
 }
